drivers: usart_linux: Fail early and release unused resources
This commit introduces early failing to csp_usart_open() in
case neither rx_callback nor return_fd are provided.
We also move the allocation of `ctx` into the if block ensuring
that a rx_callback is provided. Thus, we do not allocate and
free memory needlessly.

After usage of the pthread attributes, we destroy them to further
reduce memory leaks. In case of an error, at the moment we only
log it and do not abort, as it would not influence functionality.
If return_fd is not provided (l. 205), we do not need to explicitly
close the descriptor as rx_callback is provided and using the fd
in that case.

Signed-off-by: pr0me <g33sus@gmail.com>
Reviewed-by: Yasushi SHOJI <yashi@spacecubics.com>