csp_macro: Add CONTAINER_OF
Introduce a new macro called `CONTAINER_OF()`. This macro is designed
to return a pointer to the containing structure of the given pointer,
`ptr`. Alongside `ptr`, you are required to provide the type of the
structure and the name of the member variable within the struct.

The inclusion of an additional cast to a void pointer is necessary to
suppress warnings from Clang.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>