csp_buffer: Fix pointers in csp_buffer_clone
The `frame_begin` pointer was introduced in commit 5ee2d367ae3d5c, but
csp_buffer_clone() was not updated to reflect this change. Without
updating the pointer, the cloned buffer will always point to the
original's header[] data array, which is not our intention.

The `data_end` pointer is introduced in this csp_packet common methods
series to eventually replace packet->length and packet->frame_length.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>