csp_io: remove unused variable
This commit removes the unused variable
'csp_promisc_queue' from 'csp_io.c'.
The use of the 'extern' keyword suggests
the variable is declared somewhere as global,
but its not. The only variable with such name
is located in 'csp_promisc.c' and it is 'static'.

It seems that when moving promiscuous routine to its
own file, csp_promisc.c, the signature of csp_promisc_add()
has changed to remove the second argument, csp_promisc_queue.
The commit, 30cbebd, moved to using the new csp_promisc_add()
by introducing csp_promisc.h.