examples: Simplify device selection
Both `csp_server` and `csp_client` have lost the capability to set
multiple devices. Please refer to the `if`-block with multiple device
checks in `main()`. I plan to recover this capability in the near
future, but for now, I am simplifying them due to the fact that only
one interface can be specified.

This commit removes `can_device`, `kiss_device`, and `zmq_device`,
replacing them with a single `device_name`. It introduces
`device_type` to indicate which device is specified. With this change,
even if, for example, `libsocketcan-dev` is not installed, it won't
result in an error like:

    csp_server.c:165:25: error: ‘can_device’ undeclared

Although I can convert `if (device_type)` with `switch(device_type)`,
this `if`-block is left untouched so that I can use it when adding
multi-device capability.

In the process, this commit leverages the fact that `CSP_HAVE_*` and
`CSP_USE_*` are always defined as `1` or `0`. This allows us to
replace `#if` preprocessor macros with proper C `if`-blocks.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>