examples: csp_client: Do not use thread for client()
The function client() does not need to be run from a thread. This is a
leftover from csp_server_client. Move all functionality from client()
to main() so that we can loop over ping, reboot, and sending data to
the server in the main thread.

For test_mode, we check elapsed time in the while loop. As commented,
we don't really care if it's exactly 3 seconds or 3.99999 seconds from
the start.

Since we are removing the thread for client(), we can eliminate
task_client() and client_start. We still need csp_pthread_create for
the router for the obvious reasons.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>