doc: Fix README path
I've been testing and building doc with

    cd doc
    cmake -B builddir
    ninja -C builddir

In this case README.md is two dirs up.

However, the "official" way, or our CI's way, to build doc is

    cmake -B build-docs -S doc
    cmake --build build-docs

With this, README.md is at the parent dir.  I'd be nice if we can
write CWD independent way but this one needs a quick hot fix.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>