csp_crc32_verify: Exclude CRC size from packet length calculation
When verifying the CRC in csp_crc32_verify(), it is incorrect to
calculate the CRC for the entire packet, including the CRC itself. In
csp_crc32_append(), we generate a CRC value for the entire packet
and then append this value at the end.
This commit corrects the issue by subtracting the size of the CRC from
the frame_length, thus fixing the end address calculation.
This fix aligns with the proposal made by clythersHackers in #514.

Signed-off-by: Gaetan Perrot <gaetan.perrot@spacecubics.com>
Signed-off-by: Andrew Armitage <andrewbarmitage@gmail.com>