cmake: add CMake package export
Make allows exporting packages through its target export features:
https://cmake.org/cmake/help/latest/guide/importing-exporting/index.html#exporting-targets

This simplifies the usage of the library for other CMake projects, especially if the
library was installed using CMake.

This PR introduces the export and was done according to the CMake documentation.
It is now possible to use an installed libcsp instance by using CMake directives like
`find_package`