RDP: Fix ACK unconditionally accepted via RST + ACK
If a RST + ACK segment arrives in the OPEN state, the ACK was accepted
unconditionally.
Previously, if the received segment has a sequence number that was outside the
expected window, it would be accepted even if out of sequence.
Add a sequence number check to verify if the RST-ACK is out of sequence.

Signed-off-by: Gaetan Perrot <gaetan.perrot@spacecubics.com>