github: Checkout pushed or pull-request version of libcsp
The build-test-zephyr workflow uses `west update` to pull libcsp from
the repository, which defaults to the `develop` branch as specified in
libcsp-zephyr/west.yml.

Update build-test-zephyr.yml to fetch the newly pushed revision or the
latest commit of the current pull request instead.

Note: The `runs-on` has been updated because `ubuntu-latest` currently
points to `ubuntu-22.04`, which doesn't include `yq`.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>