github: Checkout pushed or pull-req version of libcsp
This build test uses west update to pull libcsp from the repo. This
means that it will always pull the revision defined in
libcsp-zephyr/west.yml, which is `develop`.

Instead, we want to test newlly pushed revision or the last commit of
the current pull-request.

Update build-test-zephyr.yml to do that.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>