rdp_queue: Move __csp_rdp_queue_flush before file global variables
The commit 55d443715fcc fixed a bug where the file global variable
tx_queue was accessed instead of the provided queue. This went
unnoticed because accessing tx_queue was valid, though incorrect.

To prevent this issue in the future, we are moving
__csp_rdp_queue_flush() before the file global variable
declarations. With this change, the compiler can at least issue a
warning.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>