rdp_queue: Refactor csp_rdp_queue_get for clarity
Refactored csp_rdp_queue_get by moving the else block outside of the
if condition. The csp_rdp_queue_add call is now at the end of the
loop, clearly indicating that the packet is re-added to the queue when
it doesn't match the connection.

No functionality has been changed.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>