type safety: Make data types explicit in bit modifications
Compiling for most targets, left bit shifts of "small" data types might
cause integer promotion.  On a 32 bit system, for example, left-shifting
the priority promotes it to a signed 32 bit integer.  The priority is a
2 bit value that is left-shifted to bit position 30, thereby overwriting
the sign bit.  This causes some analyzers (e.g. GCC sanitizer) to issue
an error.
The clean solution is to cast the variables to the proper lvalue data
type of the assignment *before* left-shifting.