examples: python: client: Fix options for routing table
This update fixes the usage of command line options for the routing table.

Previously, the routing table was used by default, and there was a bug in the
Python binding that would break the Python example if the rtable wasn't used.

Now, the routing table is only used when the user specifies it via a command
line option.

Signed-off-by: Gaetan Perrot <gaetan.perrot@spacecubics.com>