csp_if_eth: Add null check for CSP packet in Ethernet RX
Added a check to verify if the `csp_if_eth_pbuf_get` function returns a NULL
packet.
If the packet is NULL, the interface's drop counter is incremented, and
an error message is logged.

The function now returns `CSP_ERR_NOMEM` in case of a NULL packet, ensuring that
the error is handled properly and preventing potential issues from processing a
NULL packet.

Signed-off-by: Gaetan Perrot <gaetan.perrot@spacecubics.com>