drivers: can: zephyr: Prevent potential null pointer dereference
Rearranged the assignment of the interface name to occur after the null check on
the device pointer.

Although `device->name` wasn't necessarily accessed before the null check, it
could have been in certain scenarios, which posed a risk of null pointer
dereference.

The check is now performed first to ensure safe access.

Signed-off-by: Gaetan Perrot <gaetan.perrot@spacecubics.com>