include: csp_debug: Disable -Waddress warning for csp_assert_fail_action
Since GCC 12, the `-Waddress` warning triggers for weak function
checks, and with `-Werror` enabled, this escalates to a build
error. To maintain the stability of the libcsp v1 branch, which is in
maintenance mode, we avoid making API or ABI changes. Instead, this
commit disables the warning/error locally at the affected site.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>