github: workflows: build-test: Test with various GCC versions
Building libcsp sometimes fails with specific GCC versions due to
changes in the default warning levels. This commit adds a matrix to
test builds with GCC versions 10 through 14.

The "exclude" strategy is used instead of "include" because, despite
GitHub Actions supporting matrix expansion, I could not make it work
in this case.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>