drivers: can_zephyr: Replace CAN_MAX_DLC by CAN_MAX_DLEN
The CAN frame transmit function was using CAN_MAX_DLC to validate the data
length, which is incorrect.

CAN_MAX_DLC represents the maximum Data Length Code value, not the actual number
of bytes allowed in a frame. This could lead to memory corruption if dlc >
CAN_MAX_DLEN.

Replaced it with CAN_MAX_DLEN, which correctly represents the maximum number of
bytes that can be stored in a CAN frame.

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>
Signed-off-by: Gaetan Perrot <gaetan.perrot@spacecubics.com>